+++ src/dotnet/CarbonAware/Interfaces/ICarbonAwareDataSource.cs
@@ -1,0 +1,4 @@
+﻿namespace CarbonAware.Interfaces;
+public interface ICarbonIntensityDataSource
+{
+}
